requirements.txt
langchain
langchain_experimental
langchain_community
streamlit
openai
tiktoken
python-dotenv
pinecone-client